+++ src/common/src/test/java/org/niis/xrd4j/common/util/SOAPHelperTest.java
@@ -22,10 +22,15 @@
  */
 package org.niis.xrd4j.common.util;
 
+import jakarta.xml.soap.MessageFactory;
+import jakarta.xml.soap.Name;
 import jakarta.xml.soap.SOAPElement;
+import jakarta.xml.soap.SOAPEnvelope;
+import jakarta.xml.soap.SOAPException;
 import jakarta.xml.soap.SOAPMessage;
 import org.junit.jupiter.api.Test;
 import org.w3c.dom.NodeList;
+import org.xmlunit.assertj3.XmlAssert;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -400,5 +405,52 @@ class SOAPHelperTest {
         SOAPMessage msg = SOAPHelper.toSOAP(soapString);
 
         assertFalse(SOAPHelper.hasAttachments(msg));
+    }
+
+    @Test
+    void testMoveChildren() throws SOAPException {
+        // Envelope to be copied from
+        SOAPEnvelope envelope = MessageFactory.newInstance().createMessage().getSOAPPart().getEnvelope();
+        Name elementName = envelope.createName("TestElement", Constants.NS_XRD_PREFIX, Constants.NS_ID_URL);
+        SOAPElement testElement = envelope.getBody().addChildElement(elementName);
+        // Default Namespace
+        SOAPElement child1 = testElement.addChildElement("child1", "", "default-ns");
+        SOAPElement child11 = child1.addChildElement("child1.1", "", "default-ns");
+        child11.addChildElement("child1.1.1", "", "default-ns");
+        child11.addTextNode("Text node in child1.1.1");
+        // No namespace
+        SOAPElement child2 = testElement.addChildElement("child2");
+        SOAPElement child21 = child2.addChildElement("child2.1");
+        child21.addChildElement("child2.1.1");
+        child21.addTextNode("Text node in child2.1.1");
+        // Custom namespace, different than new parent
+        SOAPElement child3 = testElement.addChildElement("child3", "some-custom-prfx", "some-custom-ns");
+        SOAPElement child31 = child3.addChildElement("child3.1", "some-custom-prfx", "some-custom-ns");
+        child31.addChildElement("child3.1.1", "some-custom-prfx", "some-custom-ns");
+        child31.addTextNode("Text node in child3.1.1");
+
+        // Envelope to be copied to
+        SOAPEnvelope envelopeTo = MessageFactory.newInstance().createMessage().getSOAPPart().getEnvelope();
+        Name elementNameTo = envelopeTo.createName("ToElement", "new-prfx", "new-ns");
+        SOAPElement testElementTo = envelopeTo.getBody().addChildElement(elementNameTo);
+
+        SOAPHelper.moveChildren(testElement, testElementTo, true);
+
+        NodeList children = testElementTo.getChildNodes();
+
+        assertEquals("default-ns", children.item(0).getFirstChild().getNamespaceURI());
+        assertEquals(null, children.item(0).getFirstChild().getPrefix());
+        assertEquals(null, children.item(0).getFirstChild().getFirstChild().getPrefix());
+        assertEquals("new-ns", children.item(1).getFirstChild().getNamespaceURI());
+        assertEquals("new-prfx", children.item(1).getFirstChild().getPrefix());
+        assertEquals("new-prfx", children.item(1).getFirstChild().getFirstChild().getPrefix());
+        assertEquals("some-custom-ns", children.item(2).getFirstChild().getNamespaceURI());
+        assertEquals("some-custom-prfx", children.item(2).getFirstChild().getPrefix());
+        assertEquals("some-custom-prfx", children.item(2).getFirstChild().getFirstChild().getPrefix());
+
+        String toElement = SOAPHelper.toString(testElementTo);
+        XmlAssert.assertThat(toElement).withNamespaceContext(Map.of("new-prfx", "new-ns", "", "default-ns")).hasXPath("//new-prfx:ToElement/:child1/:child1.1").exist();
+        XmlAssert.assertThat(toElement).withNamespaceContext(Map.of("new-prfx", "new-ns")).hasXPath("//new-prfx:ToElement/new-prfx:child2/new-prfx:child2.1").exist();
+        XmlAssert.assertThat(toElement).withNamespaceContext(Map.of("new-prfx", "new-ns", "some-custom-prfx", "some-custom-ns")).hasXPath("//new-prfx:ToElement/some-custom-prfx:child3/some-custom-prfx:child3.1").exist();
     }
 }